requirements.txt
numpy==1.18.1
tensorflow-gpu==2.1.0
opencv-python==4.1.2.30